// src/types/dashboard/user.ts
export type User = {
  id: number;
  username: string;
  email: string;
  first_name: string;
  last_name: string;
  is_active: boolean;
  is_staff: boolean;
  roles?: string[]; // For role management
};

export type CreateUserRequest = {
  username: string;
  email: string;
  first_name: string;
  last_name: string;
  password: string;
  is_active: boolean;
  is_staff: boolean;
};

export type UpdateUserRequest = Partial<CreateUserRequest> & {
  id: number;
};